switched order of libname hints
I assume we should be searching for a lib named freeglut before a lib named glut.  In fact, should we ever be searching for a lib named glut?  Won't that pick up the regular glut rather then the freeglut version?  On windows freeglut builds a lib named "freeglut" perhaps it actually names the lib "glut" on other systems???